Factor out remaining files.
