--- frontend/lib/forms.tsx
@@ -1,153 +1,0 @@
-import React from 'react';
-import { Route } from 'react-router';
-import { WithServerFormFieldErrors, getFormErrors } from './form-errors';
-import { AppContext, AppLegacyFormSubmission } from './app-context';
-import { assertNotNull } from './util';
-import { FetchMutationInfo, createMutationSubmitHandler } from './forms-graphql';
-import { AllSessionInfo } from './queries/AllSessionInfo';
-import { getAppStaticContext } from './app-static-context';
-import { getSuccessRedirect, FormSubmitterWithoutRouter, FormSubmitterPropsWithRouter, FormSubmitterProps } from './form-submitter';
-
-
-/**
- * This component wraps a form and modifies its initial state with any information
- * passed from the server as a result of a legacy browser POST.
- */
-function LegacyFormSubmissionWrapper<FormInput, FormOutput extends WithServerFormFieldErrors>(
-  props: FormSubmitterPropsWithRouter<FormInput, FormOutput> & {
-    isSubmissionOurs: (submission: AppLegacyFormSubmission) => boolean;
-  }
-) {
-  return (
-    <AppContext.Consumer>
-      {(appCtx) => {
-        const { isSubmissionOurs, ...otherProps } = props;
-        let newProps: FormSubmitterPropsWithRouter<FormInput, FormOutput> = {
-          ...otherProps,
-          extraFields: (
-            <React.Fragment>
-              <input type="hidden" name="csrfmiddlewaretoken" value={appCtx.session.csrfToken} />
-              {props.extraFields}
-            </React.Fragment>
-          ),
-          extraFormAttributes: {
-            ...props.extraFormAttributes,
-            method: 'POST',
-            action: props.location.pathname
-          }
-        };
-        /* istanbul ignore next: this is tested by integration tests. */
-        if (appCtx.legacyFormSubmission && isSubmissionOurs(appCtx.legacyFormSubmission)) {
-          const initialState: FormInput = appCtx.legacyFormSubmission.input;
-          const output: FormOutput = appCtx.legacyFormSubmission.result;
-          const initialErrors = output.errors.length ? getFormErrors<FormInput>(output.errors) : undefined;
-          newProps = {
-            ...newProps,
-            initialState,
-            initialErrors
-          };
-          if (output.errors.length === 0) {
-            const redirect = getSuccessRedirect(newProps, initialState, output);
-            if (redirect) {
-              const appStaticCtx = assertNotNull(getAppStaticContext(props));
-              appStaticCtx.url = redirect;
-              return null;
-            }
-            // TODO: If we're still here, that means the form submission was successful.
-            // When processing forms on the client-side, we'd call the form's onSuccess
-            // handler here, but we don't want to do that here because it would likely
-            // result in a component state change, and our components are stateless
-            // during server-side rendering. So I'm not really sure what to do here.
-          }
-        }
-        return <FormSubmitterWithoutRouter {...newProps} />;
-      }}
-    </AppContext.Consumer>
-  );
-}
-
-export interface SessionUpdatingFormOutput extends WithServerFormFieldErrors {
-  session: Partial<AllSessionInfo>|null;
-}
-
-export type SessionToFormInputFn<FormInput> = ((session: AllSessionInfo) => FormInput);
-
-type SessionUpdatingFormSubmitterProps<FormInput, FormOutput extends SessionUpdatingFormOutput> = Omit<LegacyFormSubmitterProps<FormInput, FormOutput>, 'initialState'> & {
-  initialState: SessionToFormInputFn<FormInput>|FormInput;
-};
-
-/**
- * This form submitter supports a very common use case in which the GraphQL mutation,
- * when successful, simply returns a 'session' key that contains updates to the
- * session state.
- */
-export class SessionUpdatingFormSubmitter<FormInput, FormOutput extends SessionUpdatingFormOutput> extends React.Component<SessionUpdatingFormSubmitterProps<FormInput, FormOutput>> {
-  render() {
-    return (
-      <AppContext.Consumer>
-        {(appCtx) => {
-          let initialState = (typeof(this.props.initialState) === 'function')
-            // TypeScript 3.1 introduced an extremely confusing breaking change with
-            // opaque intentions [1] that broke our type narrowing, and after about an
-            // hour of trying I'm giving up and forcing a typecast. This might be easier
-            // if Typescript actually had a JSON type [2], since that's essentially what
-            // FormInput is, but otherwise I have no idea how to narrow FormInput in
-            // such a way that our conditional would result in a type narrowing instead
-            // of TypeScript 3.1's bizarre new behavior.
-            //
-            // [1] https://github.com/Microsoft/TypeScript/wiki/Breaking-Changes#narrowing-functions-now-intersects--object-and-unconstrained-generic-type-parameters
-            // [2] https://github.com/Microsoft/TypeScript/issues/1897
-            ? (this.props.initialState as SessionToFormInputFn<FormInput>)(appCtx.session)
-            : this.props.initialState;
-          return <LegacyFormSubmitter
-            {...this.props}
-            initialState={initialState}
-            onSuccess={(output) => {
-              appCtx.updateSession(assertNotNull(output.session));
-              if (this.props.onSuccess) {
-                this.props.onSuccess(output);
-              }
-            }}
-          />;
-        }}
-      </AppContext.Consumer>
-    );
-  }
-}
-
-type LegacyFormSubmitterProps<FormInput, FormOutput extends WithServerFormFieldErrors> = Omit<FormSubmitterProps<FormInput, FormOutput>, 'onSubmit'> & {
-  mutation: FetchMutationInfo<FormInput, FormOutput>
-};
-
-/** A form submitter that supports submission via legacy browser POST. */
-export class LegacyFormSubmitter<FormInput, FormOutput extends WithServerFormFieldErrors> extends React.Component<LegacyFormSubmitterProps<FormInput, FormOutput>> {
-  render() {
-    return (
-      <AppContext.Consumer>
-        {(appCtx) => {
-          return (
-            <Route render={(ctx) => {
-              const { mutation, formId, ...otherProps } = this.props;
-              const isOurs = (sub: AppLegacyFormSubmission) =>
-                sub.POST['graphql'] === mutation.graphQL &&
-                sub.POST['legacyFormId'] === formId;
-              const props: FormSubmitterProps<FormInput, FormOutput> = {
-                ...otherProps,
-                onSubmit: createMutationSubmitHandler(appCtx.fetch, mutation.fetch),
-                extraFields: (
-                  <React.Fragment>
-                    <input type="hidden" name="graphql" value={mutation.graphQL} />
-                    {formId && <input type="hidden" name="legacyFormId" value={formId}/>}
-                    {otherProps.extraFields}
-                  </React.Fragment>
-                )
-              };
-              return <LegacyFormSubmissionWrapper
-                      isSubmissionOurs={isOurs} {...props} {...ctx} />
-            }} />
-          );
-        }}
-      </AppContext.Consumer>
-    );
-  }
-}
