+++ frontend/lib/tests/forms.test.tsx
@@ -1,11 +1,14 @@
 import React from 'react';
-import { FormSubmitter, Form, BaseFormProps, FormSubmitterWithoutRouter, SessionUpdatingFormSubmitter, FormContext, BaseFormContextOptions } from '../forms';
 import { createTestGraphQlClient, pause, simpleFormErrors } from './util';
 import { shallow, mount } from 'enzyme';
 import { MemoryRouter, Route, Switch } from 'react-router';
 import { ServerFormFieldError, FormErrors } from '../form-errors';
 import { TextualFormField } from '../form-fields';
 import { AppTesterPal } from './app-tester-pal';
+import { FormContext, BaseFormContextOptions } from '../form-context';
+import { SessionUpdatingFormSubmitter } from '../session-updating-form-submitter';
+import { FormSubmitterWithoutRouter, FormSubmitter } from '../form-submitter';
+import { BaseFormProps, Form } from '../form';
 
 type MyFormOutput = {
   errors: ServerFormFieldError[],
